split c src files
