+++ tnt-kafka/producer.h
@@ -1,0 +1,61 @@
+#ifndef TNT_KAFKA_PRODUCER_H
+#define TNT_KAFKA_PRODUCER_H
+
+#include <lua.h>
+#include <lualib.h>
+#include <lauxlib.h>
+
+#include <librdkafka/rdkafka.h>
+
+////////////////////////////////////////////////////////////////////////////////////////////////////
+/**
+ * Producer
+ */
+
+typedef struct {
+    rd_kafka_topic_t **elements;
+    int32_t count;
+    int32_t capacity;
+} producer_topics_t;
+
+producer_topics_t *new_producer_topics(int32_t capacity);
+
+int add_producer_topics(producer_topics_t *topics, rd_kafka_topic_t *element);
+
+rd_kafka_topic_t *find_producer_topic_by_name(producer_topics_t *topics, const char *name);
+
+void destroy_producer_topics(producer_topics_t *topics);
+
+// Cause `rd_kafka_conf_set_events(rd_config, RD_KAFKA_EVENT_DR)` produces segfault with queue api, we are forced to
+// implement our own thread safe queue to push incoming events from callback thread to lua thread.
+typedef struct {
+    int dr_callback;
+    int err;
+} queue_element_t;
+
+queue_element_t *new_queue_element(int dr_callback, int err);
+
+void destroy_queue_element(queue_element_t *element);
+
+typedef struct {
+    rd_kafka_t        *rd_producer;
+    producer_topics_t *topics;
+    queue_t           *delivery_queue;
+    event_queues_t    *event_queues;
+} producer_t;
+
+int lua_producer_tostring(struct lua_State *L);
+
+int lua_producer_poll(struct lua_State *L);
+
+int lua_producer_msg_delivery_poll(struct lua_State *L);
+
+int lua_producer_produce(struct lua_State *L);
+
+int lua_producer_close(struct lua_State *L);
+
+int lua_producer_gc(struct lua_State *L);
+
+int lua_create_producer(struct lua_State *L);
+
+#endif //TNT_KAFKA_PRODUCER_H
