+++ tnt-kafka/queue.c
@@ -1,0 +1,101 @@
+#include <stdlib.h>
+#include <pthread.h>
+#include <unistd.h>
+
+#include <queue.h>
+
+////////////////////////////////////////////////////////////////////////////////////////////////////
+/**
+ * General thread safe queue based on licked list
+ */
+
+/**
+ * Pop without locking mutex.
+ * Caller must lock and unlock queue mutex by itself.
+ * Use with caution!
+ * @param queue
+ * @return
+ */
+void *
+queue_lockfree_pop(queue_t *queue) {
+    void *output = NULL;
+
+    if (queue->head != NULL) {
+        output = queue->head->value;
+        queue_node_t *tmp = queue->head;
+        queue->head = queue->head->next;
+        free(tmp);
+        if (queue->head == NULL) {
+            queue->tail = NULL;
+        }
+    }
+
+    return output;
+}
+
+void *
+queue_pop(queue_t *queue) {
+    pthread_mutex_lock(&queue->lock);
+
+    void *output = queue_lockfree_pop(queue);
+
+    pthread_mutex_unlock(&queue->lock);
+
+    return output;
+}
+
+int
+queue_push(queue_t *queue, void *value) {
+    if (value == NULL || queue == NULL) {
+        return -1;
+    }
+
+    pthread_mutex_lock(&queue->lock);
+
+    queue_node_t *new_node;
+    new_node = malloc(sizeof(queue_node_t));
+    if (new_node == NULL) {
+        return -1;
+    }
+
+    new_node->value = value;
+    new_node->next = NULL;
+
+    if (queue->tail != NULL) {
+        queue->tail->next = new_node;
+    }
+
+    queue->tail = new_node;
+    if (queue->head == NULL) {
+        queue->head = new_node;
+    }
+
+    pthread_mutex_unlock(&queue->lock);
+
+    return 0;
+}
+
+queue_t *
+new_queue() {
+    queue_t *queue = malloc(sizeof(queue_t));
+    if (queue == NULL) {
+        return NULL;
+    }
+
+    pthread_mutex_t lock;
+    if (pthread_mutex_init(&lock, NULL) != 0) {
+        return NULL;
+    }
+
+    queue->lock = lock;
+    queue->head = NULL;
+    queue->tail = NULL;
+
+    return queue;
+}
+
+void
+destroy_queue(queue_t *queue) {
+    pthread_mutex_destroy(&queue->lock);
+    free(queue);
+}
